Added code that finds the peaks.
It finds two kinds of peaks: heuristic and timestamp. Both use the algorithm
used in the local minima paper- find peaks in the delta_t graph of timestamps.
This code also does the same thing with the heuristic costs.

The results are extremely interesting. It seems that it will be most useful to
use a combination of heuristic and timestamp peaks to guide search. Using
their union seems to ensure that search will truly come out of the local
minimum.

This is not ensured by merely using the timestamp peaks as exit states as they
themselves lie inside the local minimum and in the absence of any preferential
treatment by A*, they will be expanded only after all the states inside the
local minimum are expanded.